Migrate content gallery to TS.
